@@ -6,10 +6,7 @@ class Row extends React.Component {
   static propTypes = {
     onClick: React.PropTypes.func,
     style: React.PropTypes.object,
-    children: React.PropTypes.oneOfType([
-      React.PropTypes.object,
-      React.PropTypes.array,
-    ]),
+    children: React.PropTypes.any,
   };
 
   getStyles() {
