@@ -32,8 +32,8 @@ class Value extends React.Component {
       float: 'left',
       width: 'auto',
       height: 'auto',
-      margin: '0',
-      padding: '0',
+      margin: '0px',
+      padding: '0px',
       fontWeight: '600',
       backgroundColor: VALUE_BACKGROUND_COLOR,
       borderRadius: VALUE_BORDER_RADIUS,
